worker_container/requirements.txt
azure-storage-blob>=12.21.0,<13.0.0
